+++ lib/elastic_autocomplete.dart
@@ -1,12 +1,13 @@
 library elastic_autocomplete;
 
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:localstorage/localstorage.dart';
 import 'package:flutter/scheduler.dart';
 import 'dart:math';
 import 'dart:convert';
 
-/// A widget wraps `RawAutocomplete`.
+/// A widget wraps [RawAutocomplete].
 class ElasticAutocomplete<T extends String> extends StatefulWidget {
   /// The user's text input is received in a field built with the
   /// [fieldViewBuilder] parameter. The options to be displayed are
@@ -219,27 +220,33 @@ class ElasticAutocompleteController<T extends String> {
   /// Store [val] to the storage unit.
   Future store(T val) async {
     var oldList = load();
-    bool append = true;
-    for (int i = 0; i < oldList.length; i++) {
-      if (oldList[i].toLowerCase() == val.toLowerCase()) {
-        append = false;
-        // update value when using case-insensitive mode
-        if (!caseSensitive) {
-          oldList[i] = val;
+    List<T> newList = <T>[];
+
+    if (latestFirst) {
+      newList.add(val);
+    }
+
+    if (caseSensitive) {
+      for (var e in oldList) {
+        if (e == val) {
+          continue;
         }
-        break;
+        newList.add(val);
+      }
+    } else {
+      for (var e in oldList) {
+        if (e.toLowerCase() == val.toLowerCase()) {
+          continue;
+        }
+        newList.add(val);
       }
     }
 
-    if (append) {
-      if (latestFirst) {
-        oldList = [val, ...oldList];
-      } else {
-        oldList.add(val);
-      }
+    if (!latestFirst) {
+      newList.add(val);
     }
 
-    _storeList(oldList);
+    _storeList(newList);
   }
 
   Future _storeList(List<T> list) async {
@@ -272,5 +279,118 @@ class ElasticAutocompleteController<T extends String> {
                   .toLowerCase()
                   .contains(textEditingValue.text.toLowerCase()));
     });
+  }
+
+  AutocompleteFieldViewBuilder fieldViewBuilder({
+    Key? key,
+    InputDecoration? decoration = const InputDecoration(),
+    TextInputType? keyboardType,
+    TextCapitalization textCapitalization = TextCapitalization.none,
+    TextInputAction? textInputAction,
+    TextStyle? style,
+    StrutStyle? strutStyle,
+    TextDirection? textDirection,
+    TextAlign textAlign = TextAlign.start,
+    TextAlignVertical? textAlignVertical,
+    bool autofocus = false,
+    bool readOnly = false,
+    ToolbarOptions? toolbarOptions,
+    bool? showCursor,
+    String obscuringCharacter = '•',
+    bool obscureText = false,
+    bool autocorrect = true,
+    SmartDashesType? smartDashesType,
+    SmartQuotesType? smartQuotesType,
+    bool enableSuggestions = true,
+    MaxLengthEnforcement? maxLengthEnforcement,
+    int? maxLines = 1,
+    int? minLines,
+    bool expands = false,
+    int? maxLength,
+    ValueChanged<String>? onChanged,
+    GestureTapCallback? onTap,
+    VoidCallback? onEditingComplete,
+    FormFieldSetter<String>? onSaved,
+    FormFieldValidator<String>? validator,
+    List<TextInputFormatter>? inputFormatters,
+    bool? enabled,
+    double cursorWidth = 2.0,
+    double? cursorHeight,
+    Radius? cursorRadius,
+    Color? cursorColor,
+    Brightness? keyboardAppearance,
+    EdgeInsets scrollPadding = const EdgeInsets.all(20.0),
+    bool? enableInteractiveSelection,
+    TextSelectionControls? selectionControls,
+    InputCounterWidgetBuilder? buildCounter,
+    ScrollPhysics? scrollPhysics,
+    Iterable<String>? autofillHints,
+    AutovalidateMode? autovalidateMode,
+    ScrollController? scrollController,
+    String? restorationId,
+    bool enableIMEPersonalizedLearning = true,
+    MouseCursor? mouseCursor,
+  }) {
+    return (
+      BuildContext context,
+      TextEditingController textEditingController,
+      FocusNode focusNode,
+      void Function() onFieldSubmitted,
+    ) {
+      return TextFormField(
+          key: key,
+          controller: textEditingController,
+          focusNode: focusNode,
+          decoration: decoration,
+          keyboardType: keyboardType,
+          textCapitalization: textCapitalization,
+          textInputAction: textInputAction,
+          style: style,
+          strutStyle: strutStyle,
+          textDirection: textDirection,
+          textAlign: textAlign,
+          textAlignVertical: textAlignVertical,
+          autofocus: autofocus,
+          readOnly: readOnly,
+          toolbarOptions: toolbarOptions,
+          showCursor: showCursor,
+          obscuringCharacter: obscuringCharacter,
+          obscureText: obscureText,
+          autocorrect: autocorrect,
+          smartDashesType: smartDashesType,
+          smartQuotesType: smartQuotesType,
+          enableSuggestions: enableSuggestions,
+          maxLengthEnforcement: maxLengthEnforcement,
+          maxLines: maxLines,
+          minLines: minLines,
+          expands: expands,
+          maxLength: maxLength,
+          onChanged: onChanged,
+          onTap: onTap,
+          onEditingComplete: onEditingComplete,
+          onFieldSubmitted: (String value) {
+            onFieldSubmitted();
+          },
+          onSaved: onSaved,
+          validator: validator,
+          inputFormatters: inputFormatters,
+          enabled: enabled,
+          cursorWidth: cursorWidth,
+          cursorHeight: cursorHeight,
+          cursorRadius: cursorRadius,
+          cursorColor: cursorColor,
+          keyboardAppearance: keyboardAppearance,
+          scrollPadding: scrollPadding,
+          enableInteractiveSelection: enableInteractiveSelection,
+          selectionControls: selectionControls,
+          buildCounter: buildCounter,
+          scrollPhysics: scrollPhysics,
+          autofillHints: autofillHints,
+          autovalidateMode: autovalidateMode,
+          scrollController: scrollController,
+          restorationId: restorationId,
+          enableIMEPersonalizedLearning: enableIMEPersonalizedLearning,
+          mouseCursor: mouseCursor);
+    };
   }
 }